// ui/src/app/api/model/transmission-context.ts
/**
 * Public Mollies Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransmissionStatusCode } from './transmission-status-code';

export interface TransmissionContext {
  id?: number;
  movieId?: number;
  torrentId?: number;
  externalId?: number;
  name?: string | null;
  magnetUri?: string | null;
  status?: TransmissionStatusCode;
}