// ui/src/app/api/model/transmission-status-code.ts
/**
 * Public Mollies Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum TransmissionStatusCode {
    Started = 'Started',
    Downloaded = 'Downloaded',
    Complete = 'Complete'
};